requirements.txt
fastapi
uvicorn
sqlmodel
openai>=1.0.0
python-dotenv
guidance
